+++ src/app/create/components/imagerequired.tsx
@@ -1,4 +1,4 @@
-import React, { useRef } from 'react';
+import React from 'react';
 import { useFieldArray, useFormContext } from 'react-hook-form';
 
 import { CircleXIcon, TriangleAlertIcon } from 'lucide-react';
@@ -23,8 +23,6 @@ function ImageRequired() {
     rules: imageValidation,
   });
 
-  const inputFile = useRef(null);
-
   const addImage = (e: React.ChangeEvent<HTMLInputElement>) => {
     if (e.target.files) {
       const uploadFile = Array.from(e.target.files);
@@ -42,7 +40,6 @@ function ImageRequired() {
           +
           <input
             className={S.imageInput}
-            ref={inputFile}
             id="image"
             type="file"
             accept="image/*"
