+++ src/app/mypage/profile/Profile.css.ts
@@ -1,6 +1,4 @@
-import { style, globalStyle, styleVariants } from '@vanilla-extract/css';
-
-import colors from '@/styles/color';
+import { style } from '@vanilla-extract/css';
 
 export const profile = style({
   display: 'flex',
@@ -14,100 +12,4 @@ export const section = style({
   display: 'flex',
   flexDirection: 'column',
   gap: '15px',
-});
-
-export const sectionTitle = style({
-  fontSize: '16px',
-  fontWeight: '800',
-});
-
-export const image = style({
-  display: 'flex',
-  justifyContent: 'center',
-});
-
-export const imageButton = style({
-  display: 'flex',
-  border: `3px solid ${colors.gray7}`,
-  borderRadius: '50%',
-  cursor: 'pointer',
-});
-
-export const profileImage = style({
-  borderRadius: '50%',
-  cursor: 'pointer',
-});
-
-export const label = style({
-  display: 'flex',
-  gap: '10px',
-  alignItems: 'center',
-});
-
-globalStyle(`${label} span`, {
-  minWidth: '120px',
-  fontWeight: '500',
-});
-
-export const inputBase = style({
-  width: '100%',
-  borderRadius: '4px',
-  padding: '12px',
-});
-
-export const input = styleVariants({
-  default: [
-    inputBase,
-    {
-      border: '1.5px solid black',
-    },
-  ],
-  disabled: [
-    inputBase,
-    {
-      border: '1.5px solid gray',
-      cursor: 'not-allowed',
-    },
-  ],
-});
-
-export const form = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '15px',
-});
-
-export const commonInputContainer = style({
-  position: 'relative',
-});
-
-export const blind = style({
-  position: 'absolute',
-  right: '10px',
-  top: '38px',
-});
-
-export const buttonContainer = style({
-  display: 'flex',
-  marginTop: '30px',
-});
-
-export const button = style({
-  padding: '15px 62px',
-  background: 'black',
-  color: 'white',
-  fontWeight: 'bold',
-  fontSize: '16px',
-  borderRadius: '4px',
-  cursor: 'pointer',
-  marginTop: '30px',
-});
-
-export const inputError = style({
-  display: 'flex',
-  gap: '4px',
-  alignItems: 'center',
-  color: 'red',
-  fontSize: '14px',
-  marginLeft: '130px',
 });
