+++ src/components/MypageProfileClientPage/components/ImageUploadModal.css.ts
@@ -1,14 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
 import colors from '@/styles/color';
-import typography from '@/styles/typo';
-
-export const image = style({
-  display: 'flex',
-  position: 'relative',
-  justifyContent: 'center',
-  marginBottom: '15px',
-});
 
 export const imageWrapper = style({
   display: 'flex',
@@ -21,6 +13,13 @@ export const imageWrapper = style({
 
 export const imageProfile = style({
   borderRadius: '50%',
+});
+
+export const image = style({
+  display: 'flex',
+  position: 'relative',
+  justifyContent: 'center',
+  marginBottom: '15px',
 });
 
 export const imageUpload = style({
@@ -47,15 +46,4 @@ export const error = style({
   fontSize: '14px',
   fontWeight: 'normal',
   marginBottom: '15px',
-});
-
-export const buttonSubmit = style({
-  ...typography.body,
-  width: '100px',
-  padding: '10px',
-  boxSizing: 'border-box',
-  borderRadius: '4px',
-  cursor: 'pointer',
-  background: colors.primary9,
-  color: colors.white,
 });
